backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-multipart==0.0.12
openpyxl==3.1.5
fpdf2==2.8.1
python-dotenv==1.0.1
slowapi==0.1.9
bcrypt==4.2.1
PyJWT==2.9.0
